Upgrade projects page test file
